+++ NEWS.md
@@ -25,6 +25,7 @@
 * `cp` and `cn` both have the default `keep_fit` argument set to TRUE.  This
   change was made to simplify the prediction methods.
 * `print.cpr_bt` returns the object invisibly, it used to return a str(x).
+* `print.cpr_cn` returns the object invisibly
 
 ## Non-User Visible Changes
 * Refactor of c++ defining basis functions, derivatives of basis functions,